Add back in short-circuit and descending sorts
Former-commit-id: ca87f5d2a0db723b3c9069b4eb2cf4bd45cb8f1f

+++ test/test_iterator.cu
@@ -262,7 +262,7 @@ void TestCounting(T base)
     //
 
     unsigned long long  max_items   = ((1ull << ((sizeof(T) * 8) - 1)) - 1);
-    unsigned long long  copy_items  = CUB_MIN(max_items - base, 100);     // potential issue with differencing overflows when T is a smaller type than can handle the offset
+    size_t  copy_items              = (size_t) CUB_MIN(max_items - base, 100);     // potential issue with differencing overflows when T is a smaller type than can handle the offset
     T                   *h_copy     = new T[copy_items];
     T                   *d_copy     = NULL;
